backend/rag/requirements.txt
PyMuPDF
python-dotenv
agraph-python
torch
transformers
pandas
numpy